+++ engines/src/engine_nc_cg_gpu.hpp
@@ -49,7 +49,7 @@ public:
   const static uint16_t N_VARS_SQ = N_VARS * N_VARS;
 
   uint8_t get_n_vars() const { return N_VARS; };
-  uint8_t get_n_ops() const { return N_OPS; };
+  uint16_t get_n_ops() const { return N_OPS; };
   uint8_t get_n_comps() const { return NC; };
   uint8_t get_z_var_idx() const { return Z_VAR; };
